r rename campaign-profiles-collection-participant-summary serializer
