+++ api/lib/infrastructure/serializers/jsonapi/campaign-profiles-collection-participant-summary-serializer.js
